Introduces GuestRepositoryTestSuite.java. Splits GuestRepositoryTest.java into two classes based on black vs. white box testing.
